Aprondar a regra de negocio da transação;

+++ src/main/java/com/example/project/service/impl/TransactionsServiceImpl.java
@@ -2,10 +2,14 @@ package com.example.project.service.impl;
 
 import com.example.project.dao.TransactionsDao;
 import com.example.project.model.Transactions;
+import com.example.project.model.User;
+import com.example.project.model.enums.RoleType;
+import com.example.project.security.AuthenticatedUser;
 import com.example.project.service.TransactionsService;
 import com.example.project.utils.Utils;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
@@ -20,23 +24,35 @@ public class TransactionsServiceImpl implements TransactionsService {
 
     @Override
     public List<Transactions> findAll() {
-        return (List<Transactions>) transactionsDao.findAll();
+        User user = AuthenticatedUser.get();
+        if (user.getProfiles().getRole().equals(RoleType.ADMIN)) {
+            return (List<Transactions>) transactionsDao.findAll();
+        }
+        return transactionsDao.findAllByUser_Id(user.getId());
     }
 
     @Override
     public Transactions findById(UUID uuid) {
-        return transactionsDao.findById(uuid).orElse(null);
+        User user = AuthenticatedUser.get();
+        Transactions transactions = transactionsDao.findById(uuid).orElse(null);
+        if (transactions != null && (user.getProfiles().getRole().equals(RoleType.ADMIN) || transactions.getUser().equals(user))) {
+            return transactions;
+        }
+        return null;
     }
 
     @Override
     public Transactions save(Transactions transactions) {
+        User user = AuthenticatedUser.get();
+        transactions.setDate(new Date());
+        transactions.setUser(user);
         return transactionsDao.save(transactions);
     }
 
     @Override
     public Transactions deleteById(UUID id) {
         Transactions transactions = transactionsDao.findById(id).orElse(null);
-        if (transactions != null) {
+        if (transactions != null && transactions.getUser().equals(AuthenticatedUser.get())) {
             transactionsDao.delete(transactions);
         }
         return transactions;
@@ -46,7 +62,7 @@ public class TransactionsServiceImpl implements TransactionsService {
     public Transactions update(Transactions transactions, UUID uuid) {
         Transactions oldTransactions = transactionsDao.findById(uuid).orElse(null);
         if (oldTransactions == null) {
-            return null;
+            return transactions;
         }
         Utils.copyNonNullProperties(oldTransactions, transactions);
         return transactionsDao.save(oldTransactions);
